decouple db models from the main class
